+++ src/Transformers/DeltaTfIdfTransformer.php
@@ -19,8 +19,8 @@ use function is_null;
  * that receive the highest boost are those whose concentration is primarily in one
  * class whereas low weighted terms are more evenly distributed among the classes.
  *
- * > **Note**: This transformer assumes that its input is made up of word frequency
- * vectors such as those created by the Word Count Vectorizer.
+ * > **Note**: Delta TF-IDF Transformer assumes that its inputs are made up of token
+ * frequency vectors such as those created by the Word Count Vectorizer.
  *
  * References:
  * [1] J. Martineau et al. (2009). Delta TFIDF: An Improved Feature Space for
@@ -32,8 +32,16 @@ use function is_null;
  * @package     Rubix/ML
  * @author      Andrew DalPino
  */
-class DeltaTfIdfTransformer implements Elastic
+class DeltaTfIdfTransformer implements Transformer, Stateful, Elastic
 {
+    /**
+     * The amount of additive (Laplace) smoothing to add to the inverse document
+     * frequencies (IDFs).
+     *
+     * @var float
+     */
+    protected $smoothing;
+
     /**
      * The class specific term frequencies of each word i.e. the number of
      * times a word appears in the context of a class label.
@@ -81,6 +89,19 @@ class DeltaTfIdfTransformer implements Elastic
     protected $entropies;
 
     /**
+     * @param float $smoothing
+     */
+    public function __construct(float $smoothing = 1.0)
+    {
+        if ($smoothing <= 0.0) {
+            throw new InvalidArgumentException('Smoothing must be'
+                . " greater than 0, $smoothing given.");
+        }
+
+        $this->smoothing = $smoothing;
+    }
+
+    /**
      * Return the data types that this transformer is compatible with.
      *
      * @return \Rubix\ML\DataType[]
@@ -103,6 +124,16 @@ class DeltaTfIdfTransformer implements Elastic
     }
 
     /**
+     * Return the document frequencies calculated during fitting.
+     *
+     * @return int[]|null
+     */
+    public function dfs() : ?array
+    {
+        return $this->dfs;
+    }
+
+    /**
      * Fit the transformer to the dataset.
      *
      * @param \Rubix\ML\Datasets\Dataset $dataset
@@ -117,11 +148,11 @@ class DeltaTfIdfTransformer implements Elastic
 
         $classes = $dataset->possibleOutcomes();
 
-        $ones = array_fill(0, $dataset->numColumns(), 1);
+        $zeros = array_fill(0, $dataset->numColumns(), 0);
 
-        $this->tfs = array_fill_keys($classes, $ones);
-        $this->dfs = $this->totals = $ones;
-        $this->n = 1;
+        $this->tfs = array_fill_keys($classes, $zeros);
+        $this->dfs = $this->totals = $zeros;
+        $this->n = 0;
 
         $this->update($dataset);
     }
@@ -138,7 +169,7 @@ class DeltaTfIdfTransformer implements Elastic
                 . ' labeled training set.');
         }
 
-        SamplesAreCompatibleWithTransformer::check($dataset, $this);
+        SamplesAreCompatibleWithTransformer::with($dataset, $this)->check();
 
         if (is_null($this->tfs) or is_null($this->dfs)) {
             $this->fit($dataset);
@@ -166,17 +197,19 @@ class DeltaTfIdfTransformer implements Elastic
 
         $this->n += $dataset->numRows();
 
+        $nHat = $this->n + $this->smoothing;
+
         $idfs = [];
 
         foreach ($this->dfs as $df) {
-            $idfs[] = 1.0 + log($this->n / $df);
+            $idfs[] = 1.0 + log($nHat / ($df + $this->smoothing));
         }
 
         $entropies = array_fill(0, count($this->totals), 0.0);
 
         foreach ($this->tfs as $tfs) {
             foreach ($tfs as $column => $tf) {
-                $delta = $tf / $this->totals[$column];
+                $delta = ($tf + $this->smoothing) / ($this->totals[$column] + $this->smoothing);
 
                 $entropies[$column] += -$delta * log($delta);
             }
@@ -215,6 +248,6 @@ class DeltaTfIdfTransformer implements Elastic
      */
     public function __toString() : string
     {
-        return 'Delta TF-IDF Transformer';
+        return "Delta TF-IDF Transformer (smoothing: {$this->smoothing})";
     }
 }