+++ CHANGELOG.md
@@ -1,3 +1,15 @@
-- Unreleased
+- 0.2.0-beta1
+    - Implemented Token Hashing Vectorizer transformer
+    - Move Recursive Feature Eliminator (RFE) to main repo
+    - Added Laplace smoothing to Delta TF-IDF Transformer
+
+- 0.1.0-beta
+    - Add Recursive Feature Eliminator feature selector
+    - Implement BM25 TF-IDF Transformer
+    - Add Delta TF-IDF Transformer
     - Added Lambda function Transformer
     - All objects implement Stringable interface
+    - Added Gower nan-safe distance kernel
+    - Added ISRU and ISRLU activation functions
+    - Added Alpha Dropout hidden layer
+    - Added Model Orchestra meta-estimator